Add files via upload
[B]Krogsbell Version 2019-06-05[/B]
New Addon name: Danish Auto Netradio ;-)
Purpose of this version: Play constantly a preferred DR station. Switch to other DR stations if
interesting programs start. Switch to backup station if unwanted program is started on preferred station.
Added EPG for DR channels.
Logic to select wanted and unwanted programs with preferred station and backup station.
In EPG view wanted programs are marked with green text and unwanted programs are marked with red text.
After start, the preferred station will start.
If unwanted programs on the preferred station, the backup station is started.
If wanted programs on any DR station, the wanted program is started and when finished, the preferred station is started.
xmltodict.py added from github release 12
Added icons to stations in EPG views
All times internal are utc, the EPG view shows local time.
You select which P4 station to use in the settings
Changed use of service.py to enable change in settings not to be overwritten